refactor: move set_temporal_aggregation to a new rule
