+++ rules/build_sector.smk
@@ -1408,7 +1408,6 @@ def input_heat_source_power(w):
 
 rule prepare_sector_network:
     params:
-        time_resolution=config_provider("clustering", "temporal", "resolution_sector"),
         co2_budget=config_provider("co2_budget"),
         conventional_carriers=config_provider(
             "existing_capacities", "conventional_carriers"
@@ -1444,9 +1443,6 @@ rule prepare_sector_network:
         unpack(input_heat_source_power),
         **rules.cluster_gas_network.output,
         **rules.build_gas_input_locations.output,
-        snapshot_weightings=resources(
-            "snapshot_weightings_base_s_{clusters}_elec_{opts}_{sector_opts}.csv"
-        ),
         retro_cost=lambda w: (
             resources("retro_cost_base_s_{clusters}.csv")
             if config_provider("sector", "retrofitting", "retro_endogen")(w)
@@ -1564,7 +1560,7 @@ rule prepare_sector_network:
         ),
     output:
         resources(
-            "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}.nc"
+            "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}_h.nc"
         ),
     threads: 1
     resources:
@@ -1579,3 +1575,32 @@ rule prepare_sector_network:
         )
     script:
         "../scripts/prepare_sector_network.py"
+
+
+rule temporal_aggregation:
+    params:
+        time_resolution=config_provider("clustering", "temporal", "resolution_sector"),
+    input:
+        network=resources(
+            "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}_h.nc"
+        ),
+        snapshot_weightings=resources(
+            "snapshot_weightings_base_s_{clusters}_elec_{opts}_{sector_opts}.csv"
+        ),
+    output:
+        resources(
+            "networks/base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}.nc"
+        ),
+    threads: 1
+    resources:
+        mem_mb=5000,
+    log:
+        logs(
+            "temporal_aggregation_base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}.log"
+        ),
+    benchmark:
+        benchmarks(
+            "temporal_aggregation_base_s_{clusters}_{opts}_{sector_opts}_{planning_horizons}"
+        )
+    script:
+        "../scripts/temporal_aggregation.py"
