+++ scripts/prepare_sector_network.py
@@ -5716,58 +5716,6 @@ def cluster_heat_buses(n):
         n.add(c.name, df.loc[to_add].index, **df.loc[to_add])
 
 
-def set_temporal_aggregation(n, resolution, snapshot_weightings):
-    """
-    Aggregate time-varying data to the given snapshots.
-    """
-    if not resolution:
-        logger.info("No temporal aggregation. Using native resolution.")
-        return n
-    elif "sn" in resolution.lower():
-        # Representative snapshots are dealt with directly
-        sn = int(resolution[:-2])
-        logger.info("Use every %s snapshot as representative", sn)
-        n.set_snapshots(n.snapshots[::sn])
-        n.snapshot_weightings *= sn
-        return n
-    else:
-        # Otherwise, use the provided snapshots
-        snapshot_weightings = pd.read_csv(
-            snapshot_weightings, index_col=0, parse_dates=True
-        )
-
-        # Define a series used for aggregation, mapping each hour in
-        # n.snapshots to the closest previous timestep in
-        # snapshot_weightings.index
-        aggregation_map = (
-            pd.Series(
-                snapshot_weightings.index.get_indexer(n.snapshots), index=n.snapshots
-            )
-            .replace(-1, np.nan)
-            .ffill()
-            .astype(int)
-            .map(lambda i: snapshot_weightings.index[i])
-        )
-
-        m = n.copy(snapshots=[])
-        m.set_snapshots(snapshot_weightings.index)
-        m.snapshot_weightings = snapshot_weightings
-
-        # Aggregation all time-varying data.
-        for c in n.iterate_components():
-            pnl = getattr(m, c.list_name + "_t")
-            for k, df in c.pnl.items():
-                if not df.empty:
-                    if c.list_name == "stores" and k == "e_max_pu":
-                        pnl[k] = df.groupby(aggregation_map).min()
-                    elif c.list_name == "stores" and k == "e_min_pu":
-                        pnl[k] = df.groupby(aggregation_map).max()
-                    else:
-                        pnl[k] = df.groupby(aggregation_map).mean()
-
-        return m
-
-
 def lossy_bidirectional_links(n, carrier, efficiencies={}):
     """Split bidirectional links into two unidirectional links to include transmission losses."""
 
@@ -6425,10 +6373,6 @@ if __name__ == "__main__":
 
     if options["allam_cycle_gas"]:
         add_allam_gas(n, costs, pop_layout=pop_layout, spatial=spatial)
-
-    n = set_temporal_aggregation(
-        n, snakemake.params.time_resolution, snakemake.input.snapshot_weightings
-    )
 
     co2_budget = snakemake.params.co2_budget
     if isinstance(co2_budget, str) and co2_budget.startswith("cb"):
